requirements.txt
urllib3==1.23
xlrd==1.1.0
requests==2.20.0
xlwt==1.3.0
pymssql==2.1.4
colorama==0.3.9
PySnooper==0.0.37
setuptools==40.5.0
Flask==1.0.2
six==1.11.0
click==6.7
PyYAML==5.1
coverage==4.5.2
coveralls==1.5.1
pylint==2.3.1